+++ jannovar-cli/src/main/java/de/charite/compbio/jannovar/cmd/annotate_vcf/AnnotateVCFCommand.java
@@ -1,11 +1,7 @@
 package de.charite.compbio.jannovar.cmd.annotate_vcf;
 
-import htsjdk.variant.variantcontext.VariantContext;
-import htsjdk.variant.vcf.VCFFileReader;
-import htsjdk.variant.vcf.VCFHeader;
-
 import java.io.File;
-import java.util.function.Supplier;
+import java.io.IOException;
 import java.util.stream.Stream;
 
 import org.apache.commons.cli.ParseException;
@@ -16,9 +12,20 @@ import de.charite.compbio.jannovar.ProgressReporter;
 import de.charite.compbio.jannovar.cmd.CommandLineParsingException;
 import de.charite.compbio.jannovar.cmd.HelpRequestedException;
 import de.charite.compbio.jannovar.cmd.JannovarAnnotationCommand;
+import de.charite.compbio.jannovar.mendel.filter.ConsumerProcessor;
+import de.charite.compbio.jannovar.mendel.filter.CoordinateSortingChecker;
+import de.charite.compbio.jannovar.mendel.filter.GeneWiseMendelianAnnotationProcessor;
+import de.charite.compbio.jannovar.mendel.filter.VariantContextProcessor;
+import de.charite.compbio.jannovar.pedigree.PedFileContents;
+import de.charite.compbio.jannovar.pedigree.PedFileReader;
+import de.charite.compbio.jannovar.pedigree.PedParseException;
+import de.charite.compbio.jannovar.pedigree.Pedigree;
 import de.charite.compbio.jannovar.vardbs.base.DBAnnotationOptions;
 import de.charite.compbio.jannovar.vardbs.facade.DBVariantContextAnnotator;
 import de.charite.compbio.jannovar.vardbs.facade.DBVariantContextAnnotatorFactory;
+import htsjdk.variant.variantcontext.VariantContext;
+import htsjdk.variant.vcf.VCFFileReader;
+import htsjdk.variant.vcf.VCFHeader;
 
 /**
  * Run annotation steps (read in VCF, write out VCF or Jannovar file format).
@@ -101,18 +108,44 @@ public class AnnotateVCFCommand extends JannovarAnnotationCommand {
 
 				// Write result to output file
 				try (AnnotatedVCFWriter writer = new AnnotatedVCFWriter(refDict, vcfHeader, chromosomeMap, vcfPath,
-						options, args)) {
-					stream.forEachOrdered(writer::put);
+						options, args); VariantContextProcessor sink = buildMendelianProcessors(writer);) {
+					stream.forEachOrdered(sink::put);
 
 					System.err.println("Wrote annotations to \"" + writer.getOutFileName() + "\"");
 					final long endTime = System.nanoTime();
 					System.err.println(String.format("Annotation and writing took %.2f sec.",
 							(endTime - startTime) / 1000.0 / 1000.0 / 1000.0));
+				} catch (IOException e) {
+					throw new JannovarException("Problem opening file", e);
 				}
 			}
 		}
 		if (progressReporter != null)
 			progressReporter.stop();
+	}
+
+	/**
+	 * Construct the mendelian inheritance annotation processors
+	 * 
+	 * @param sink
+	 *            The place to put put the VariantContext to after filtration
+	 * @throws IOException
+	 *             in case of problems with opening the pedigree file
+	 * @throws PedParseException
+	 *             in the case of problems with parsing pedigrees
+	 */
+	private VariantContextProcessor buildMendelianProcessors(AnnotatedVCFWriter writer)
+			throws PedParseException, IOException {
+		if (options.pathPedFile != null) {
+			final PedFileReader pedReader = new PedFileReader(new File(options.pathPedFile));
+			final PedFileContents pedContents = pedReader.read();
+			final Pedigree pedigree = new Pedigree(pedContents, pedContents.getIndividuals().get(0).getPedigree());
+			final GeneWiseMendelianAnnotationProcessor mendelProcessor = new GeneWiseMendelianAnnotationProcessor(
+					pedigree, jannovarData, vc -> writer.put(vc));
+			return new CoordinateSortingChecker(mendelProcessor);
+		} else {
+			return new ConsumerProcessor(vc -> writer.put(vc));
+		}
 	}
 
 	@Override