+++ jannovar-htsjdk/src/main/java/de/charite/compbio/jannovar/mendel/filter/GeneIntervalEndExtractor.java
@@ -1,4 +1,4 @@
-package de.charite.compbio.jannovar.filter;
+package de.charite.compbio.jannovar.mendel.filter;
 
 import de.charite.compbio.jannovar.impl.intervals.IntervalEndExtractor;
 
@@ -6,6 +6,7 @@ import de.charite.compbio.jannovar.impl.intervals.IntervalEndExtractor;
  * Extraction of interval ends of {@link Gene} objects.
  */
 class GeneIntervalEndExtractor implements IntervalEndExtractor<Gene> {
+
 	public int getBegin(Gene gene) {
 		return gene.getRegion().getBeginPos();
 	}
@@ -13,4 +14,5 @@ class GeneIntervalEndExtractor implements IntervalEndExtractor<Gene> {
 	public int getEnd(Gene gene) {
 		return gene.getRegion().getEndPos();
 	}
+
 }